Updates due to HSD Commands integration

+++ src/main/java/com/st/STWINBoard_Gui/Utils/SensorViewAdapter.kt
@@ -17,7 +17,7 @@ import com.st.clab.stwin.gui.R
 typealias OnSubSensorEnableStatusChange = (sensor:Sensor,subSensor: SubSensorDescriptor, newState:Boolean)->Unit
 typealias OnSubSensorODRChange = (sensor:Sensor,subSensor: SubSensorDescriptor, newOdrValue:Double)->Unit
 typealias OnSubSensorFullScaleChange = (sensor:Sensor,subSensor: SubSensorDescriptor, newFSValue:Double)->Unit
-typealias OnSubSensorSampleChange = (sensor:Sensor,subSensor: SubSensorDescriptor, newSampleValue:Double)->Unit
+typealias OnSubSensorSampleChange = (sensor:Sensor,subSensor: SubSensorDescriptor, newSampleValue:Int)->Unit
 
 class SensorViewAdapter(//Activity Context
         private val onSubSubSensorEnableStatusChange: OnSubSensorEnableStatusChange,
